+++ src/MooVC.Architecture/Ddd/AggregateRootExtensions.Coordinate.cs
@@ -1,7 +1,6 @@
 ﻿namespace MooVC.Architecture.Ddd
 {
     using System;
-    using System.Threading.Tasks;
     using static MooVC.Architecture.Ddd.Resources;
     using static MooVC.Ensure;
 
@@ -14,18 +13,6 @@
             aggregate
                 .GetType()
                 .Coordinate(
-                    aggregate.Id,
-                    operation,
-                    timeout: timeout);
-        }
-
-        public static async Task CoordinateAsync(this AggregateRoot aggregate, Func<Task> operation, TimeSpan? timeout = default)
-        {
-            ArgumentNotNull(aggregate, nameof(aggregate), AggregateRootExtensionsCoordinateAggregateRequired);
-
-            await aggregate
-                .GetType()
-                .CoordinateAsync(
                     aggregate.Id,
                     operation,
                     timeout: timeout);
