+++ src/MooVC.Architecture/Ddd/ReferenceExtensions.Coordinate.cs
@@ -1,7 +1,6 @@
 ﻿namespace MooVC.Architecture.Ddd
 {
     using System;
-    using System.Threading.Tasks;
     using static MooVC.Architecture.Ddd.Resources;
     using static MooVC.Ensure;
 
@@ -9,23 +8,11 @@
     {
         public static void Coordinate(this Reference reference, Action operation, TimeSpan? timeout = default)
         {
-            ArgumentNotNull(reference, nameof(reference), ReferenceExtensionsCoordinateReferenceeRequired);
+            ArgumentNotNull(reference, nameof(reference), ReferenceExtensionsCoordinateReferenceRequired);
 
             reference
                 .Type
                 .Coordinate(
-                    reference.Id,
-                    operation,
-                    timeout: timeout);
-        }
-
-        public static async Task CoordinateAsync(this Reference reference, Func<Task> operation, TimeSpan? timeout = default)
-        {
-            ArgumentNotNull(reference, nameof(reference), ReferenceExtensionsCoordinateReferenceeRequired);
-
-            await reference
-                .Type
-                .CoordinateAsync(
                     reference.Id,
                     operation,
                     timeout: timeout);
