+++ src/MooVC.Architecture/MessageExtensions.Coordinate.cs
@@ -1,7 +1,6 @@
 ﻿namespace MooVC.Architecture
 {
     using System;
-    using System.Threading.Tasks;
     using static MooVC.Architecture.Resources;
     using static MooVC.Ensure;
 
@@ -14,18 +13,6 @@
             message
                 .GetType()
                 .Coordinate(
-                    message.Id,
-                    operation,
-                    timeout: timeout);
-        }
-
-        public static async Task CoordinateAsync(this Message message, Func<Task> operation, TimeSpan? timeout = default)
-        {
-            ArgumentNotNull(message, nameof(message), MessageExtensionsCoordinateMessageRequired);
-
-            await message
-                .GetType()
-                .CoordinateAsync(
                     message.Id,
                     operation,
                     timeout: timeout);
